GP2-1850: Minor refactor to move key-pages and test api out of `search` into `activitystream`
This creates better separation between the activitystream app (sending content to AS) and search (getting content from AS)
