+++ activitystream/templates/search-key-pages.json
